example_pipeline/pyproject.toml
[project]
name = "example_pipeline"
version = "0.0.1"
authors = [{ name = "james@posit.co" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
    "requests>=2.28.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
]

[project.scripts]
main = "example_pipeline.main:main"
fetch_weather = "example_pipeline.weather_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 125
